Many changes detailed below:
* Working on the footer, made a standard footer for all the websites,
created a snippet for it on vim, so can be called with a shortcut.
The footer is responsive and it changes substantially between layouts,
it is not strictly necessary as it looked good before, but now looks much
better.
Included a snippet of both html and scss code for the footer, and
footer variables too. I'm not uploading this file as it is not
part of the site or development tools.

* Added smooth scrolling and managed to work fine, but still it's not
seen on my old iphone, for some reason.

Next steps are
* to make the site responsive and extend from phones to desktop.
* fix the scrolling.
* Upload a couple of posts (taking advantage of the snippets all this
should be a fast process.)