The mobile version is ready. There are ton of improvements to be added but overall it's functional, looks decent and it's probably gonna be easy to extend.
The main thing to improve is the card, it still looks
rather ugly.

The changes on this commit are summarized as:
* created toggler.js containing all toggle-like scripts.
* Try to create an ordered meaning for the links, particularly the yellow color, which implies a link-like behaviour even if, under the hood, the element is not a link.
Example: the card header with a yellow chevron.
* Updated the grid because the row gaps were not
what I was looking for. Then I created rows that act as row gaps, but then I can control the height.